# [FIX] mail_message: improve perf of message_fetch_failed