@robodoo 2024-01-29
[Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/151668).