@Aurelienvd 2024-02-01
@adda-odoo Hello, I didn't review the PR yet but could you add more info to the benchmark ? Like the scanned tables' cardinality and the recordset cardinality if there is any ? Thanks !