@seb-odoo 2024-02-02
Hello, thank you for the investigation.
Note that the missing indexes have been added starting from v16 at https://github.com/odoo/odoo/pull/83451

v14 is no longer supported so this PR shouldn't be merged. Best case scenario would be to fix only in v15, but is it really worth it?

I don't think it's possible to get a satisfactory result in stable without the missing indexes? 70% reduction is not really big, it's not even an order of magnitude. 20s might be reduced to 6s but that's still really bad.